V0.2.3 Day 13
Cleans up the movement of hillary.  Gives the developer more control.
Goes from using a move to command to checking donalds position and
moving in the correct x direction.  Working on adding in a jump system
and its own gravity for hillary.  Is also ready to take collision inputs
from a collision with donald.